+++ packages/grid/src/features/standard-table/config/StandardTableColumnConfig.ts
@@ -9,7 +9,37 @@ export type StandardTableColumnConfig<
   TItem,
   TItemValue
 > = StandardTableColumnOptions<TItem, TItemValue> &
+  StandardTableColumnOptionsWithNoGroups &
   ItemValueResolver<TItem, TItemValue>;
+
+export type StandardTableColumnConfigWithGroups<
+  TItem,
+  TItemValue
+> = StandardTableColumnOptions<TItem, TItemValue> &
+  ItemValueResolver<TItem, TItemValue>;
+
+export interface StandardTableColumnOptionsWithNoGroups {
+  /**
+   * Enable sticky behaviour to the left make elements scroll in behind this column.
+   * If neither left nor right is specified, it defaults to left: 0px.
+   */
+  sticky?: boolean;
+
+  /**
+   * Set a custom z index
+   */
+  zIndex?: number;
+
+  /**
+   * Offset column from left (ex if we have multiple sticky columns)
+   */
+  left?: string;
+
+  /**
+   * Offset column from right (ex if we have multiple sticky columns)
+   */
+  right?: string;
+}
 
 export interface StandardTableColumnOptions<TItem, TItemValue> {
   /**
@@ -106,22 +136,6 @@ export interface StandardTableColumnOptions<TItem, TItemValue> {
     | "isEditable"
     | "onChange"
   >;
-
-  /**
-   * Enable sticky behaviour to the left
-   * make elements scroll in behind this column
-   */
-  sticky?: boolean;
-
-  /**
-   * Set a custom z index
-   */
-  zIndex?: number;
-
-  /**
-   * Offset column from left (ex if we have multiple sticky columns)
-   */
-  left?: string;
 
   /**
    * The icon variant to use when displaying sort order.