+++ packages/grid/src/features/standard-table/context/StickyColumnOffsetContext.ts
@@ -1,9 +1,0 @@
-import { createContext, useContext } from "react";
-import { OffsetPerColumn } from "../util/StickyColumnGroupOffsetCalculator";
-
-export const StickyColumnOffsetContext = createContext<OffsetPerColumn<string>>(
-  {}
-);
-
-export const useStickyColumnOffsetContext = () =>
-  useContext(StickyColumnOffsetContext);