+++ packages/grid/src/features/standard-table/features/sticky-columns/types.ts
@@ -1,0 +1,10 @@
+export type StickyPropsPerColumn<TColumn extends string> = Record<
+  TColumn,
+  StickyCellProps
+>;
+
+export interface StickyCellProps {
+  left: string | undefined;
+  right: string | undefined;
+  sticky: boolean;
+}